+++ leetcode/Linear Search in C++.cpp
