+++ leetcode/Minimum Interval to Include Each Query.cpp
