+++ leetcode/N-queens2.cpp
