+++ leetcode/binarySearch.cpp
