+++ leetcode/rectanglepattern.cpp
